+++ src/collective/contentsections/contents/__init__.py
@@ -1,0 +1,2 @@
+from .location.content import ILocation
+from .location.content import Location